+++ src/main/java/racingcar/move/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.move;
 
 public interface MoveStrategy {
     boolean isMovable();
