+++ src/main/java/racingcar/move/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.move;
 
 public interface NumberGenerator {
     int generateNumber();
